requirements-core.txt
# JARVIS Core Service Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
aioredis==2.0.1
chromadb==0.4.18
httpx==0.25.2
python-multipart==0.0.6
python-dotenv==1.0.0
prometheus-client==0.19.0
loguru==0.7.2
pyyaml==6.0.1
